feat: 重构 RouteActivity 为 Fragment 容器
