+++ core/chatai/app/src/main/java/me/rerere/rikkahub/RouteFragment.kt
@@ -1,0 +1,698 @@
+package me.rerere.rikkahub
+
+import android.content.Intent
+import android.os.Build
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.SharedTransitionLayout
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.scaleIn
+import androidx.compose.animation.scaleOut
+import androidx.compose.animation.slideInHorizontally
+import androidx.compose.animation.slideOutHorizontally
+import androidx.compose.animation.togetherWith
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.SideEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.ComposeView
+import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.testTagsAsResourceId
+import androidx.compose.ui.unit.dp
+import androidx.core.net.toUri
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.viewmodel.navigation3.rememberViewModelStoreNavEntryDecorator
+import androidx.navigation3.runtime.NavKey
+import androidx.navigation3.runtime.entryProvider
+import androidx.navigation3.runtime.rememberNavBackStack
+import androidx.navigation3.runtime.rememberSaveableStateHolderNavEntryDecorator
+import androidx.navigation3.ui.NavDisplay
+import coil3.ImageLoader
+import coil3.compose.setSingletonImageLoaderFactory
+import coil3.gif.AnimatedImageDecoder
+import coil3.gif.GifDecoder
+import coil3.network.cachecontrol.CacheControlCacheStrategy
+import coil3.network.okhttp.OkHttpNetworkFetcherFactory
+import coil3.request.crossfade
+import coil3.svg.SvgDecoder
+import com.dokar.sonner.Toaster
+import com.dokar.sonner.rememberToasterState
+import kotlinx.serialization.Serializable
+import me.rerere.highlight.Highlighter
+import me.rerere.highlight.LocalHighlighter
+import me.rerere.rikkahub.data.datastore.SettingsStore
+import me.rerere.rikkahub.data.db.DatabaseMigrationTracker
+import me.rerere.rikkahub.data.db.MigrationState
+import me.rerere.rikkahub.data.event.AppEvent
+import me.rerere.rikkahub.data.event.AppEventBus
+import me.rerere.rikkahub.ui.components.ui.TTSController
+import me.rerere.rikkahub.ui.context.LocalASRState
+import me.rerere.rikkahub.ui.context.LocalNavController
+import me.rerere.rikkahub.ui.context.LocalSettings
+import me.rerere.rikkahub.ui.context.LocalSharedTransitionScope
+import me.rerere.rikkahub.ui.context.LocalTTSState
+import me.rerere.rikkahub.ui.context.LocalToaster
+import me.rerere.rikkahub.ui.context.Navigator
+import me.rerere.rikkahub.ui.hooks.readBooleanPreference
+import me.rerere.rikkahub.ui.hooks.readStringPreference
+import me.rerere.rikkahub.ui.hooks.rememberCustomAsrState
+import me.rerere.rikkahub.ui.hooks.rememberCustomTtsState
+import me.rerere.rikkahub.ui.pages.assistant.AssistantPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantBasicPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantDetailPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantExtensionsPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantLocalToolPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantMcpPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantMemoryPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantPromptPage
+import me.rerere.rikkahub.ui.pages.assistant.detail.AssistantRequestPage
+import me.rerere.rikkahub.ui.pages.backup.BackupPage
+import me.rerere.rikkahub.ui.pages.chat.ChatPage
+import me.rerere.rikkahub.ui.pages.debug.DebugPage
+import me.rerere.rikkahub.ui.pages.extensions.ExtensionsPage
+import me.rerere.rikkahub.ui.pages.extensions.PromptPage
+import me.rerere.rikkahub.ui.pages.extensions.QuickMessagesPage
+import me.rerere.rikkahub.ui.pages.extensions.skills.SkillDetailPage
+import me.rerere.rikkahub.ui.pages.extensions.skills.SkillsPage
+import me.rerere.rikkahub.ui.pages.extensions.workspace.WorkspaceDetailPage
+import me.rerere.rikkahub.ui.pages.extensions.workspace.WorkspacePage
+import me.rerere.rikkahub.ui.pages.extensions.workspace.WorkspaceTerminalPage
+import me.rerere.rikkahub.ui.pages.favorite.FavoritePage
+import me.rerere.rikkahub.ui.pages.history.HistoryPage
+import me.rerere.rikkahub.ui.pages.imggen.ImageGenPage
+import me.rerere.rikkahub.ui.pages.log.LogPage
+import me.rerere.rikkahub.ui.pages.search.SearchPage
+import me.rerere.rikkahub.ui.pages.setting.SettingAboutPage
+import me.rerere.rikkahub.ui.pages.setting.SettingDonatePage
+import me.rerere.rikkahub.ui.pages.setting.SettingFilesPage
+import me.rerere.rikkahub.ui.pages.setting.SettingMcpPage
+import me.rerere.rikkahub.ui.pages.setting.SettingModelPage
+import me.rerere.rikkahub.ui.pages.setting.SettingPage
+import me.rerere.rikkahub.ui.pages.setting.SettingPreferencesGeneralPage
+import me.rerere.rikkahub.ui.pages.setting.SettingPreferencesNotificationPage
+import me.rerere.rikkahub.ui.pages.setting.SettingPreferencesPage
+import me.rerere.rikkahub.ui.pages.setting.SettingPreferencesThemePage
+import me.rerere.rikkahub.ui.pages.setting.SettingPreferencesUIPage
+import me.rerere.rikkahub.ui.pages.setting.SettingProviderDetailPage
+import me.rerere.rikkahub.ui.pages.setting.SettingProviderPage
+import me.rerere.rikkahub.ui.pages.setting.SettingSearchDetailPage
+import me.rerere.rikkahub.ui.pages.setting.SettingSearchPage
+import me.rerere.rikkahub.ui.pages.setting.SettingSpeechPage
+import me.rerere.rikkahub.ui.pages.setting.SettingThemePage
+import me.rerere.rikkahub.ui.pages.setting.SettingWebPage
+import me.rerere.rikkahub.ui.pages.share.handler.ShareHandlerPage
+import me.rerere.rikkahub.ui.pages.stats.StatsPage
+import me.rerere.rikkahub.ui.pages.translator.TranslatorPage
+import me.rerere.rikkahub.ui.pages.webview.WebViewPage
+import me.rerere.rikkahub.ui.theme.LocalDarkMode
+import me.rerere.rikkahub.ui.theme.RikkahubTheme
+import okhttp3.OkHttpClient
+import org.koin.android.ext.android.inject
+import org.koin.compose.koinInject
+import kotlin.uuid.Uuid
+
+class RouteFragment : Fragment() {
+
+    private val highlighter by inject<Highlighter>()
+    private val okHttpClient by inject<OkHttpClient>()
+    private val settingsStore by inject<SettingsStore>()
+
+    private var navStack: MutableList<NavKey>? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        return ComposeView(requireContext()).apply {
+            setViewCompositionStrategy(
+                ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed
+            )
+            setContent {
+                RikkahubTheme {
+                    setSingletonImageLoaderFactory { context ->
+                        ImageLoader.Builder(context)
+                            .crossfade(true)
+                            .components {
+                                add(
+                                    OkHttpNetworkFetcherFactory(
+                                        callFactory = { okHttpClient },
+                                        cacheStrategy = { CacheControlCacheStrategy() },
+                                    )
+                                )
+                                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+                                    add(AnimatedImageDecoder.Factory())
+                                } else {
+                                    add(GifDecoder.Factory())
+                                }
+                                add(SvgDecoder.Factory(scaleToDensity = true))
+                            }
+                            .build()
+                    }
+                    AppRoutes()
+                }
+            }
+        }
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        (activity as? RouteActivity)?.registerConversationNavigator { id ->
+            navStack?.add(Screen.Chat(id))
+        }
+    }
+
+    override fun onDestroyView() {
+        (activity as? RouteActivity)?.registerConversationNavigator(null)
+        super.onDestroyView()
+    }
+
+    @Composable
+    private fun ShareHandler(backStack: MutableList<NavKey>) {
+        val shareIntent = remember {
+            Intent().apply {
+                action = activity?.intent?.action
+                putExtra(Intent.EXTRA_TEXT, activity?.intent?.getStringExtra(Intent.EXTRA_TEXT))
+                putExtra(Intent.EXTRA_STREAM, activity?.intent?.getStringExtra(Intent.EXTRA_STREAM))
+                putExtra(
+                    Intent.EXTRA_PROCESS_TEXT,
+                    activity?.intent?.getCharSequenceExtra(Intent.EXTRA_PROCESS_TEXT)
+                )
+            }
+        }
+
+        LaunchedEffect(backStack) {
+            when (shareIntent.action) {
+                Intent.ACTION_SEND -> {
+                    val text = shareIntent.getStringExtra(Intent.EXTRA_TEXT) ?: ""
+                    val imageUri = shareIntent.getStringExtra(Intent.EXTRA_STREAM)
+                    backStack.add(Screen.ShareHandler(text, imageUri))
+                }
+
+                Intent.ACTION_PROCESS_TEXT -> {
+                    val text = shareIntent.getCharSequenceExtra(Intent.EXTRA_PROCESS_TEXT)?.toString() ?: ""
+                    backStack.add(Screen.ShareHandler(text, null))
+                }
+            }
+        }
+    }
+
+    @OptIn(ExperimentalComposeUiApi::class)
+    @Composable
+    fun AppRoutes() {
+        val toastState = rememberToasterState()
+        val settings by settingsStore.settingsFlow.collectAsStateWithLifecycle()
+        val tts = rememberCustomTtsState()
+        val asr = rememberCustomAsrState()
+        val eventBus = koinInject<AppEventBus>()
+        LaunchedEffect(tts) {
+            eventBus.events.collect { event ->
+                when (event) {
+                    is AppEvent.Speak -> tts.speak(event.text)
+                }
+            }
+        }
+        val migrationState by DatabaseMigrationTracker.state.collectAsStateWithLifecycle()
+
+        val startScreen = Screen.Chat(
+            id = if (readBooleanPreference("create_new_conversation_on_start", true)) {
+                Uuid.random().toString()
+            } else {
+                readStringPreference(
+                    "lastConversationId",
+                    Uuid.random().toString()
+                ) ?: Uuid.random().toString()
+            }
+        )
+
+        val backStack = rememberNavBackStack(startScreen)
+        SideEffect { this@RouteFragment.navStack = backStack }
+
+        ShareHandler(backStack)
+
+        SharedTransitionLayout {
+            CompositionLocalProvider(
+                LocalNavController provides Navigator(backStack),
+                LocalSharedTransitionScope provides this,
+                LocalSettings provides settings,
+                LocalHighlighter provides highlighter,
+                LocalToaster provides toastState,
+                LocalTTSState provides tts,
+                LocalASRState provides asr,
+            ) {
+                Toaster(
+                    state = toastState,
+                    darkTheme = LocalDarkMode.current,
+                    richColors = true,
+                    alignment = Alignment.TopCenter,
+                    showCloseButton = true,
+                )
+                TTSController()
+                Box(
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .semantics { testTagsAsResourceId = true }
+                        .background(MaterialTheme.colorScheme.background)
+                ) {
+                    NavDisplay(
+                        backStack = backStack,
+                        entryDecorators = listOf(
+                            rememberSaveableStateHolderNavEntryDecorator(),
+                            rememberViewModelStoreNavEntryDecorator(),
+                        ),
+                        modifier = Modifier.fillMaxSize(),
+                        onBack = { backStack.removeLastOrNull() },
+                        transitionSpec = {
+                            if (backStack.size == 1) fadeIn() togetherWith fadeOut()
+                            else {
+                                slideInHorizontally { it } togetherWith
+                                    slideOutHorizontally { -it / 2 } + scaleOut(targetScale = 0.7f) + fadeOut()
+                            }
+                        },
+                        popTransitionSpec = {
+                            slideInHorizontally { -it / 2 } + scaleIn(initialScale = 0.7f) + fadeIn() togetherWith
+                                slideOutHorizontally { it }
+                        },
+                        predictivePopTransitionSpec = {
+                            slideInHorizontally { -it / 2 } + scaleIn(initialScale = 0.7f) + fadeIn() togetherWith
+                                slideOutHorizontally { it }
+                        },
+                        entryProvider = entryProvider {
+                            entry<Screen.Chat>(
+                                metadata = NavDisplay.transitionSpec { fadeIn() togetherWith fadeOut() }
+                                        + NavDisplay.popTransitionSpec { fadeIn() togetherWith fadeOut() }
+                            ) { key ->
+                                ChatPage(
+                                    id = Uuid.parse(key.id),
+                                    text = key.text,
+                                    files = key.files.map { it.toUri() },
+                                    nodeId = key.nodeId?.let { Uuid.parse(it) }
+                                )
+                            }
+
+                            entry<Screen.ShareHandler> { key ->
+                                ShareHandlerPage(
+                                    text = key.text,
+                                    image = key.streamUri
+                                )
+                            }
+
+                            entry<Screen.History> {
+                                HistoryPage()
+                            }
+
+                            entry<Screen.Favorite> {
+                                FavoritePage()
+                            }
+
+                            entry<Screen.Assistant> {
+                                AssistantPage()
+                            }
+
+                            entry<Screen.AssistantDetail> { key ->
+                                AssistantDetailPage(key.id)
+                            }
+
+                            entry<Screen.AssistantBasic> { key ->
+                                AssistantBasicPage(key.id)
+                            }
+
+                            entry<Screen.AssistantPrompt> { key ->
+                                AssistantPromptPage(key.id)
+                            }
+
+                            entry<Screen.AssistantMemory> { key ->
+                                AssistantMemoryPage(key.id)
+                            }
+
+                            entry<Screen.AssistantRequest> { key ->
+                                AssistantRequestPage(key.id)
+                            }
+
+                            entry<Screen.AssistantMcp> { key ->
+                                AssistantMcpPage(key.id)
+                            }
+
+                            entry<Screen.AssistantLocalTool> { key ->
+                                AssistantLocalToolPage(key.id)
+                            }
+
+                            entry<Screen.AssistantInjections> { key ->
+                                AssistantExtensionsPage(key.id)
+                            }
+
+                            entry<Screen.Translator> {
+                                TranslatorPage()
+                            }
+
+                            entry<Screen.Setting> {
+                                SettingPage()
+                            }
+
+                            entry<Screen.Backup> {
+                                BackupPage()
+                            }
+
+                            entry<Screen.ImageGen> {
+                                ImageGenPage()
+                            }
+
+                            entry<Screen.WebView> { key ->
+                                WebViewPage(key.url, key.content)
+                            }
+
+                            entry<Screen.SettingTheme> {
+                                SettingThemePage()
+                            }
+
+                            entry<Screen.SettingPreferences> {
+                                SettingPreferencesPage()
+                            }
+
+                            entry<Screen.SettingPreferencesTheme> {
+                                SettingPreferencesThemePage()
+                            }
+
+                            entry<Screen.SettingPreferencesNotification> {
+                                SettingPreferencesNotificationPage()
+                            }
+
+                            entry<Screen.SettingPreferencesGeneral> {
+                                SettingPreferencesGeneralPage()
+                            }
+
+                            entry<Screen.SettingPreferencesUI> {
+                                SettingPreferencesUIPage()
+                            }
+
+                            entry<Screen.SettingProvider> {
+                                SettingProviderPage()
+                            }
+
+                            entry<Screen.SettingProviderDetail> { key ->
+                                val id = Uuid.parse(key.providerId)
+                                SettingProviderDetailPage(id = id)
+                            }
+
+                            entry<Screen.SettingModels> {
+                                SettingModelPage()
+                            }
+
+                            entry<Screen.SettingAbout> {
+                                SettingAboutPage()
+                            }
+
+                            entry<Screen.SettingSearch> {
+                                SettingSearchPage()
+                            }
+
+                            entry<Screen.SettingSearchDetail> { key ->
+                                val id = Uuid.parse(key.serviceId)
+                                SettingSearchDetailPage(id)
+                            }
+
+                            entry<Screen.SettingSpeech> {
+                                SettingSpeechPage()
+                            }
+
+                            entry<Screen.SettingMcp> {
+                                SettingMcpPage()
+                            }
+
+                            entry<Screen.SettingDonate> {
+                                SettingDonatePage()
+                            }
+
+                            entry<Screen.SettingFiles> {
+                                SettingFilesPage()
+                            }
+
+                            entry<Screen.SettingWeb> {
+                                SettingWebPage()
+                            }
+
+                            entry<Screen.Debug> {
+                                DebugPage()
+                            }
+
+                            entry<Screen.Log> {
+                                LogPage()
+                            }
+
+                            entry<Screen.Extensions> {
+                                ExtensionsPage()
+                            }
+
+                            entry<Screen.QuickMessages> {
+                                QuickMessagesPage()
+                            }
+
+                            entry<Screen.Prompts> {
+                                PromptPage()
+                            }
+
+                            entry<Screen.Skills> {
+                                SkillsPage()
+                            }
+
+                            entry<Screen.Workspaces> {
+                                WorkspacePage()
+                            }
+
+                            entry<Screen.WorkspaceDetail> { key ->
+                                WorkspaceDetailPage(key.id)
+                            }
+
+                            entry<Screen.WorkspaceTerminal> { key ->
+                                WorkspaceTerminalPage(key.id)
+                            }
+
+                            entry<Screen.SkillDetail> { key ->
+                                SkillDetailPage(skillName = key.skillName)
+                            }
+
+                            entry<Screen.MessageSearch> {
+                                SearchPage()
+                            }
+
+                            entry<Screen.Stats> {
+                                StatsPage()
+                            }
+                        }
+                    )
+                    if (BuildConfig.DEBUG) {
+                        Text(
+                            text = "[开发模式]",
+                            modifier = Modifier
+                                .align(Alignment.TopCenter)
+                                .padding(top = 4.dp),
+                            style = MaterialTheme.typography.labelSmall,
+                            color = MaterialTheme.colorScheme.error.copy(alpha = 0.7f)
+                        )
+                    }
+                    AnimatedVisibility(
+                        visible = migrationState is MigrationState.Migrating,
+                        enter = fadeIn(),
+                        exit = fadeOut(),
+                        modifier = Modifier.fillMaxSize()
+                    ) {
+                        val state = migrationState as? MigrationState.Migrating
+                        Box(
+                            modifier = Modifier
+                                .fillMaxSize()
+                                .background(MaterialTheme.colorScheme.surface.copy(alpha = 0.95f)),
+                            contentAlignment = Alignment.Center
+                        ) {
+                            Column(
+                                horizontalAlignment = Alignment.CenterHorizontally,
+                                verticalArrangement = Arrangement.spacedBy(16.dp)
+                            ) {
+                                CircularProgressIndicator()
+                                Text(
+                                    text = stringResource(R.string.db_migrating),
+                                    style = MaterialTheme.typography.bodyLarge
+                                )
+                                if (state != null) {
+                                    Text(
+                                        text = "v${state.from} → v${state.to}",
+                                        style = MaterialTheme.typography.bodySmall,
+                                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+sealed interface Screen : NavKey {
+    @Serializable
+    data class Chat(
+        val id: String,
+        val text: String? = null,
+        val files: List<String> = emptyList(),
+        val nodeId: String? = null
+    ) : Screen
+
+    @Serializable
+    data class ShareHandler(val text: String, val streamUri: String? = null) : Screen
+
+    @Serializable
+    data object History : Screen
+
+    @Serializable
+    data object Favorite : Screen
+
+    @Serializable
+    data object Assistant : Screen
+
+    @Serializable
+    data class AssistantDetail(val id: String) : Screen
+
+    @Serializable
+    data class AssistantBasic(val id: String) : Screen
+
+    @Serializable
+    data class AssistantPrompt(val id: String) : Screen
+
+    @Serializable
+    data class AssistantMemory(val id: String) : Screen
+
+    @Serializable
+    data class AssistantRequest(val id: String) : Screen
+
+    @Serializable
+    data class AssistantMcp(val id: String) : Screen
+
+    @Serializable
+    data class AssistantLocalTool(val id: String) : Screen
+
+    @Serializable
+    data class AssistantInjections(val id: String) : Screen
+
+    @Serializable
+    data object Translator : Screen
+
+    @Serializable
+    data object Setting : Screen
+
+    @Serializable
+    data object Backup : Screen
+
+    @Serializable
+    data object ImageGen : Screen
+
+    @Serializable
+    data class WebView(val url: String = "", val content: String = "") : Screen
+
+    @Serializable
+    data object SettingTheme : Screen
+
+    @Serializable
+    data object SettingPreferences : Screen
+
+    @Serializable
+    data object SettingPreferencesTheme : Screen
+
+    @Serializable
+    data object SettingPreferencesNotification : Screen
+
+    @Serializable
+    data object SettingPreferencesGeneral : Screen
+
+    @Serializable
+    data object SettingPreferencesUI : Screen
+
+    @Serializable
+    data object SettingProvider : Screen
+
+    @Serializable
+    data class SettingProviderDetail(val providerId: String) : Screen
+
+    @Serializable
+    data object SettingModels : Screen
+
+    @Serializable
+    data object SettingAbout : Screen
+
+    @Serializable
+    data object SettingSearch : Screen
+
+    @Serializable
+    data class SettingSearchDetail(val serviceId: String) : Screen
+
+    @Serializable
+    data object SettingSpeech : Screen
+
+    @Serializable
+    data object SettingMcp : Screen
+
+    @Serializable
+    data object SettingDonate : Screen
+
+    @Serializable
+    data object SettingFiles : Screen
+
+    @Serializable
+    data object SettingWeb : Screen
+
+    @Serializable
+    data object Debug : Screen
+
+    @Serializable
+    data object Log : Screen
+
+    @Serializable
+    data object Extensions : Screen
+
+    @Serializable
+    data object QuickMessages : Screen
+
+    @Serializable
+    data object Prompts : Screen
+
+    @Serializable
+    data object Skills : Screen
+
+    @Serializable
+    data object Workspaces : Screen
+
+    @Serializable
+    data class WorkspaceDetail(val id: String) : Screen
+
+    @Serializable
+    data class WorkspaceTerminal(val id: String) : Screen
+
+    @Serializable
+    data class SkillDetail(val skillName: String) : Screen
+
+    @Serializable
+    data object MessageSearch : Screen
+
+    @Serializable
+    data object Stats : Screen
+}
